Various changes - Added function to trim the whitespace around numbers - Don't print the half-parsed number when there is an error - Widened the integer type to 64 bits - Changed the parsing algorithm to be more cache-friendly

+++ base-conv.h
@@ -1,8 +1,9 @@
 #ifndef BASE_CONV_H
 #define BASE_CONV_H
+#include <stdint.h>
 
-int parse_number(const char *string, unsigned char base, unsigned *number); 
-char* serialize_number(unsigned int number, unsigned char base, char *buf); 
+int parse_number(const char *string, unsigned char base, uint64_t *number); 
+char* serialize_number(uint64_t number, unsigned char base, char *buf); 
 
 char char_to_digit(char chr); 
 char digit_to_char(char digit);
